pyproject.toml
[tool.poetry]
name = "mitsuba"
version = "0.1.0"
description = "a discord bot"
authors = ["alostsock <hi@alostsock.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.1.1"
schema = "^0.7.5"
sqlalchemy = "^2.0.4"
alembic = "^1.9.4"
asyncpg = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.1.0"
black = "^23.1.0"
isort = "^5.12.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
